move params.js to core folder
