+++ bigbluebutton-tests/puppeteer/core/page.js
@@ -7,7 +7,7 @@ const moment = require('moment');
 const path = require('path');
 const PuppeteerVideoRecorder = require('puppeteer-video-recorder');
 const helper = require('./helper');
-const params = require('../params');
+const params = require('./params');
 const { ELEMENT_WAIT_TIME } = require('./constants');
 const { getElementLength } = require('./util');
 const e = require('./elements');
